FIX: correct data augmentation to be coupled to the subject results and not the subjects anatomy
The code was previously doing it the other way around. The results were coupled to the subjects anatomy, i.e. different dti maps were presented with the warped to the subject's anatomy, but were expected to give the same result. The way I want to do it is to have the same dti mapped to different subjects anatomy. But, that dti map regardless of where it is warped is expected to give the same result. This way the model might learn the DTI features irrespective of the morpheological features.